+++ src/be_lexer.c
@@ -42,7 +42,7 @@ void be_lexerror(blexer *lexer, const char *msg)
 
 static void keyword_registe(bvm *vm)
 {
-    size_t i, n = array_count(kwords_tab);
+    int i, n = (int)array_count(kwords_tab);
     for (i = KeyIf; i < n; ++i) {
         bstring *s = be_newstr(vm, kwords_tab[i]);
         be_gc_fix(vm, gc_object(s));
@@ -52,7 +52,7 @@ static void keyword_registe(bvm *vm)
 
 static void keyword_unregiste(bvm *vm)
 {
-    size_t i, n = array_count(kwords_tab);
+    int i, n = (int)array_count(kwords_tab);
     for (i = KeyIf; i < n; ++i) {
         bstring *s = be_newstr(vm, kwords_tab[i]);
         be_gc_unfix(vm, gc_object(s));
